refactored code and removed unnecessary files
