+++ src/components/StartSection/StartSection.module.css
@@ -5,7 +5,7 @@
 }
 
 .contentWrapper {
-    background-image: url('../../assets/background/backgroundGrid.png');
+    background-image: url('../../assets/background/noise_transparent.svg');
 }
 
 .contentContainer {
